requirements.txt
flask~=2.0.0
kafka-python~=2.0.0
psycopg2-binary~=2.8.0
numpy~=1.22